Cleanup with actor class
