+++ gym_carla/envs/sensors/sensors.py
@@ -5,79 +5,6 @@ import logging
 from queue import Queue, Empty
 
 
-class SensorManager:
-    def __init__(self):
-        self._sensors = []
-        self._sensors_queue = {}  # Queue()  # {name: Sensor object}
-        # self._data_buffers = #Queue()
-        self._queue_timeout = 10
-        self._sensor_timeout = 2
-
-        self._event_sensors_queue = {}  # Queue()
-        # self._event_data_buffers = Queue()
-
-    @property
-    def sensors(self):
-        #sensors = {name: sensor for name, sensor in self._sensors_queue.items()}
-        #sensors.update({name: sensor for name, sensor in self._event_sensors_queue.items()})
-        return self._sensors
-
-    def reset(self):
-        for sensor in self.sensors:
-            sensor.close()
-        self._sensors_queue = {}
-        self._event_sensors_queue = {}#Queue()
-
-        # self._data_buffers = Queue()
-        # self._event_data_buffers = Queue()
-
-    def close(self):
-        self.reset()
-        #self._sensors_queue = None
-        #self._event_sensors_queue = None
-        # self._data_buffers = None
-        # self._event_data_buffers = None
-
-    def register(self, sensor):
-        """Adds a specific sensor to the class"""
-        self._sensors.append(sensor)
-        if sensor.is_event_sensor:
-            self._event_sensors_queue[sensor.name] = sensor.data_queue
-        else:
-            self._sensors_queue[sensor.name] = sensor.data_queue
-
-    def get_data(self, w_frame):
-        data_all = {}
-        if not self._sensors_queue:
-            for sensor_name, sensor_queue in self._sensor_queues.items():
-                while True:
-                    # Ensure that data is always from the same world frame w_frame
-                    frame, data = sensor_queue.get(timeout=2)
-                    if frame == w_frame:
-                        logging.debug(f"Queue {sensor_name}: {frame} == {w_frame}")
-                        return data
-                    else:
-                        logging.warning(f"Queue {sensor_name}: {frame} != {w_frame}")
-                data_all.update({sensor_name: (frame, data)})
-
-        if not self._event_sensors_queue:
-            for sensor_name, sensor_queue in self._event_sensors_queue.items():
-                while True:
-                    try:
-                        # Ensure that data is always from the same world frame w_frame
-                        frame, data = sensor_queue.get_nowait()
-                        if frame == w_frame:
-                            logging.debug(f"Queue {sensor_name}: {frame} == {w_frame}")
-                            return data
-                        else:
-                            logging.warning(f"Queue {sensor_name}: {frame} != {w_frame}")
-                    except Empty:
-                        frame = w_frame
-                        data = None
-                data_all.update({sensor_name: (frame, data)})
-        return data_all
-
-
 class Sensor(ABC):
     def __init__(self, name, params, manager, parent, bp, transform):
         # Use a queue for thread safe access to the data
